reset-password-ui

+++ server/src/controllers/user.ts
@@ -1,11 +1,6 @@
-import {
-  CreateUser,
-  ReVerifyEmailRequest,
-  UpdatePassword,
-  VerifyEmailRequest,
-} from "#/@types/user";
-import EmailVerificationToken from "#/models/emailVerificationToken";
-import PasswordResetToken from "#/models/passwordResetToken";
+import { RequestHandler } from "express";
+
+import { CreateUser, VerifyEmailRequest } from "#/@types/user";
 import User from "#/models/user";
 import { generateToken } from "#/utils/helper";
 import {
@@ -13,17 +8,19 @@ import {
   sendPassResetSuccessMail,
   sendVerificationMail,
 } from "#/utils/mail";
-import { RequestHandler } from "express";
+import EmailVerificationToken from "#/models/emailVerificationToken";
+import PasswordResetToken from "#/models/passwordResetToken";
 import { isValidObjectId } from "mongoose";
 import crypto from "crypto";
 import { PASSWORD_RESET_LINK } from "#/utils/variables";
 
 export const create: RequestHandler = async (req: CreateUser, res) => {
   const { email, password, name } = req.body;
-  const user = await User.create({ email, password, name });
 
+  const user = await User.create({ name, email, password });
+
+  // send verification email
   const token = generateToken();
-
   await EmailVerificationToken.create({
     owner: user._id,
     token,
@@ -45,30 +42,27 @@ export const verifyEmail: RequestHandler = async (
   });
 
   if (!verificationToken)
-    return res.status(403).json({ message: "Invalid token" });
+    return res.status(403).json({ error: "Invalid token!" });
 
   const matched = await verificationToken.compareToken(token);
+  if (!matched) return res.status(403).json({ error: "Invalid token!" });
 
-  if (!matched) return res.status(403).json({ message: "Invalid token" });
-
-  await User.findByIdAndUpdate(userId, { verified: true });
-
+  await User.findByIdAndUpdate(userId, {
+    verified: true,
+  });
   await EmailVerificationToken.findByIdAndDelete(verificationToken._id);
 
-  res.json({ message: "Your email is verified" });
+  res.json({ message: "Your email is verified." });
 };
 
-export const sendReVerificationToken: RequestHandler = async (
-  req: ReVerifyEmailRequest,
-  res
-) => {
+export const sendReVerificationToken: RequestHandler = async (req, res) => {
   const { userId } = req.body;
 
   if (!isValidObjectId(userId))
-    return res.status(403).json({ message: "Invalid request!" });
+    return res.status(403).json({ error: "Invalid request!" });
 
   const user = await User.findById(userId);
-  if (!user) return res.status(403).json({ message: "Invalid request!" });
+  if (!user) return res.status(403).json({ error: "Invalid request!" });
 
   await EmailVerificationToken.findOneAndDelete({
     owner: userId,
@@ -82,20 +76,19 @@ export const sendReVerificationToken: RequestHandler = async (
   });
 
   sendVerificationMail(token, {
-    name: user.name,
-    email: user.email,
-    userId: user._id.toString(),
+    name: user?.name,
+    email: user?.email,
+    userId: user?._id.toString(),
   });
 
-  res.json({ message: "Please check your mail" });
+  res.json({ message: "Please check you mail." });
 };
 
 export const generateForgetPasswordLink: RequestHandler = async (req, res) => {
   const { email } = req.body;
 
   const user = await User.findOne({ email });
-
-  if (!user) return res.status(404).json({ message: "Account not found!" });
+  if (!user) return res.status(404).json({ error: "Account not found!" });
 
   await PasswordResetToken.findOneAndDelete({
     owner: user._id,
@@ -112,39 +105,31 @@ export const generateForgetPasswordLink: RequestHandler = async (req, res) => {
 
   sendForgetPasswordLink({ email: user.email, link: resetLink });
 
-  res.json({ message: "Check your registered mail." });
+  res.json({ message: "Check you registered mail." });
 };
 
 export const grantValid: RequestHandler = async (req, res) => {
   res.json({ valid: true });
 };
 
-export const updatePassword: RequestHandler = async (
-  req: UpdatePassword,
-  res
-) => {
+export const updatePassword: RequestHandler = async (req, res) => {
   const { password, userId } = req.body;
 
   const user = await User.findById(userId);
-
-  if (!user) return res.status(403).json({ message: "Unauthorized access!" });
+  if (!user) return res.status(403).json({ error: "Unauthorized access!" });
 
   const matched = await user.comparePassword(password);
-
   if (matched)
     return res
       .status(422)
-      .json({ message: "The new password must be different!" });
+      .json({ error: "The new password must be different!" });
 
   user.password = password;
-
   await user.save();
 
-  await PasswordResetToken.findOneAndDelete({
-    owner: user._id,
-  });
+  await PasswordResetToken.findOneAndDelete({ owner: user._id });
+  // send the success email
 
   sendPassResetSuccessMail(user.name, user.email);
-
-  res.json({ message: "Password updated successfully." });
+  res.json({ message: "Password resets successfully." });
 };
